+++ test/fuzz/accounts/AccountV1/UpgradeAccount.fuzz.t.sol
@@ -21,7 +21,7 @@ contract UpgradeAccount_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
 
     struct Checks {
         bool isCreditorSet;
-        uint16 accountVersion;
+        uint88 accountVersion;
         address baseCurrency;
         address owner;
         address liquidator;
@@ -68,7 +68,7 @@ contract UpgradeAccount_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     function testFuzz_Revert_upgradeAccount_Reentered(
         address newImplementation,
         address newRegistry,
-        uint16 newVersion,
+        uint88 newVersion,
         bytes calldata data
     ) public {
         // Reentrancy guard is in locked state.
@@ -84,7 +84,7 @@ contract UpgradeAccount_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     function testFuzz_Revert_upgradeAccount_NonFactory(
         address newImplementation,
         address newRegistry,
-        uint16 newVersion,
+        uint88 newVersion,
         address nonFactory,
         bytes calldata data
     ) public {
@@ -100,7 +100,7 @@ contract UpgradeAccount_AccountV1_Fuzz_Test is AccountV1_Fuzz_Test {
     function testFuzz_Revert_upgradeAccount_InvalidAccountVersion(
         address newImplementation,
         address newRegistry,
-        uint16 newVersion,
+        uint88 newVersion,
         bytes calldata data
     ) public {
         // Given: Creditor is set.
